refactor(db): DB actionsをrepositoryへ分割する
#79に対応し、既存のDbActions公開面は維持したままusers/events/matches/match watchers/auth/static data/guildsのrepositoryへ内部実装を分割する。あわせてroute側のdbActions依存型を必要なmethodだけへ狭め、マージ済み#76/#78をTASKS.mdへ同期する。
