+++ TASKS.md
@@ -52,8 +52,8 @@ ADTeemoの詳細な未実装・改善作業はGitHub Issuesを正として追跡
   - [ ] [#72 Riot静的データ取得をBackend APIへ集約する](https://github.com/akgm3i/ADTeemo/issues/72)
   - [ ] [#73 OP.GG試合詳細の解決・保存をBackend APIへ集約する](https://github.com/akgm3i/ADTeemo/issues/73)
 - [ ] [#71 Backend/Botの依存注入と試合監視モジュールを段階的に整理する](https://github.com/akgm3i/ADTeemo/issues/71)
-  - [ ] [#76 DB factoryとDB actions factoryを導入する](https://github.com/akgm3i/ADTeemo/issues/76)
-  - [ ] [#78 createAppとroute factoryでBackend依存を注入する](https://github.com/akgm3i/ADTeemo/issues/78)
+  - [x] [#76 DB factoryとDB actions factoryを導入する](https://github.com/akgm3i/ADTeemo/issues/76)
+  - [x] [#78 createAppとroute factoryでBackend依存を注入する](https://github.com/akgm3i/ADTeemo/issues/78)
   - [ ] [#79 DB actionsをドメイン別repositoryへ分割する](https://github.com/akgm3i/ADTeemo/issues/79)
   - [ ] [#75 API client factoryと共通transport処理を導入する](https://github.com/akgm3i/ADTeemo/issues/75)
   - [ ] [#74 API clientをBackend resource別clientへ分割する](https://github.com/akgm3i/ADTeemo/issues/74)
